requirements.txt
httpx[http2]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
tenacity==8.3.0
pyyaml==6.0.1
rapidfuzz==3.6.1
feedparser==6.0.8
tabulate==0.9.0
pytest==7.4.0
openai
